move game fetching into hook
