+++ src/App.jsx
@@ -11,6 +11,7 @@ import { backdateTimestamp, formatDate, formatTimestamp } from './services/dateH
 import { ChannelRadioButton } from './components/ChannelRadioButton';
 import './App.css';
 import { Filters } from './components/Filters';
+import { useFetchGames } from './hooks/useFetchGames';
 
 // `b` is returned from ts_headline when a match is found
 const ALLOWED_TAGS = ['b'];
@@ -512,7 +513,6 @@ const App = () => {
     const [sortOrder, setSortOrder] = useState("default");
     const [feedbackModalOpen, setFeedbackModalOpen] = useState(false);
     const [submittingFeedback, setSubmittingFeedback] = useState(false);
-    const [games, setGames] = useState([]);
     const [selectedGame, setSelectedGame] = useState("all");
     
     const strict = false;
@@ -532,72 +532,7 @@ const App = () => {
         viewportMeta.content = 'width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no';
     }, []);
 
-    useEffect(() => {
-        const fetchGames = async () => {
-            try {
-                // Try multiple path configurations to handle potential Render.com path issues
-                const pathsToTry = [
-                    '/api/games',
-                    '/games',
-                    '/app/api/games'
-                ];
-
-                let gamesData = null;
-                let failureMessages = [];
-
-                // Try each path until one works
-                for (const path of pathsToTry) {
-                    try {
-                        console.log(`Trying to fetch games from: ${path}`);
-                        const response = await fetch(path, {
-                            cache: 'no-store', // Disable caching
-                            headers: {
-                                'Cache-Control': 'no-cache, no-store, must-revalidate',
-                                'Pragma': 'no-cache',
-                                'Expires': '0'
-                            }
-                        });
-
-                        // Check if we got a valid response
-                        if (response.ok) {
-                            const data = await response.json();
-                            if (data && data.games && Array.isArray(data.games)) {
-                                gamesData = data;
-                                console.log(`Successfully fetched games from ${path}`);
-                                break;
-                            } else {
-                                console.log(`Response from ${path} didn't contain valid games data`);
-                                failureMessages.push(`Invalid data from ${path}`);
-                            }
-                        } else {
-                            const errorMsg = `Failed to fetch games from ${path}: ${response.status}`;
-                            console.log(errorMsg);
-                            failureMessages.push(errorMsg);
-                        }
-                    } catch (pathError) {
-                        const errorMsg = `Error fetching games from ${path}: ${pathError.message}`;
-                        console.log(errorMsg);
-                        failureMessages.push(errorMsg);
-                    }
-                }
-
-                // If we got data from any of the paths, use it
-                if (gamesData && gamesData.games) {
-                    setGames(gamesData.games);
-                } else {
-                    // When no paths worked, set empty array but log detailed error info
-                    console.error('All paths failed. Details:', failureMessages.join('; '));
-                    console.log('Database connection issue detected - using empty games array');
-                    setGames([]);
-                }
-            } catch (error) {
-                console.error('Error fetching games:', error);
-                // Set empty array as fallback
-                setGames([]);
-            }
-        };
-        fetchGames();
-    }, []);
+    const games = useFetchGames();
 
     const handleChannelChange = (channelId) => {
         setSelectedChannel(channelId);
